requirements.txt
reportlab
fastapi
uvicorn
mysql-connector-python
pydantic
python-dotenv
sqlalchemy
databases
bcrypt
passlib[bcrypt]
python-jose[cryptography]
email-validator
PyJWT
jinja2
python-multipart